Make --precision f32 the default
The only argument against it was that it lost single-threaded.  That was
the AVX-512 scatter (ca9111d), and with that gone `:f32` wins at `-t 1` on
both development hosts: 1.206x on fitzroy (14.33 -> 11.88 s) and 1.190x on
the laptop, at the riptide bench config.  It halves the bandwidth of every
bulk array the transform and the metric stream through, and the win grows
with thread count.

Everything reported stays Float64; only the profile stage narrows.
`--precision f64` still selects the old behaviour.

The machine-precision pins now pass `precision=PIN_PRECISION` EXPLICITLY
rather than inheriting it, which is the same pattern the Float32
interpolation weights established: run at the new default they would fold
the profile stage's ~1e-7 into bounds whose job is to catch chunking,
tabulation and harmonic-gather bugs at 1e-8.  The two transpose testsets
pin `:f64` outright, since one of them exists to exercise the widening nest
that the `:f32` default dispatches away from.

568/568 tests pass.  All three Python-oracle checks pass unchanged: kernel
3.847e-16, profiles 2.131e-16, metric 1.393e-16.

Adds `bench/insitu_phase.jl`, a one-arm in-situ wall-clock + phase-share
report, for sweeping a `const` that cannot be varied in-process.

Also records a sibling-repo trap: the metric check failed at 5.694e-02
until `../coherent_search` was pulled forward past the snr1 metric change.
The crossval script prints the path it used, and that path looked correct
the whole time -- a stale checkout is indistinguishable from a current one
in that line.

+++ src/search.jl
@@ -140,7 +140,7 @@ Base.@kwdef struct SearchParams
                                        # Float32 batched gate lands within this of
                                        # `threshold` (see `boxcar_metrics!`)
     decimations::Vector{Int} = [1]  # harmonic-decimation factors k (see decimation_design.md)
-    precision::Symbol = :f64   # profile-stage element type — :f64 or :f32 (see `proftype`)
+    precision::Symbol = :f32   # profile-stage element type — :f32 or :f64 (see `proftype`)
 end
 
 """
